create compression module to dry out_dir, overwrite, git_hook_command module
